fix purge speed calculation
The previous implementation divided the flow rate by the arbitrary
number 5.0. Obviously this cannot be correct as the flow rate depends
on the amount of material transported through the extruder and as
such depends on the filament diameter.

In this implementation we retrieve the filament diameter from the
printer configuration, calculate the area of the used filament and
divide the flow rate by this area to retrieve the correct value.